Move literal visitors into separate file
